+++ CHANGELOG.md
@@ -15,6 +15,12 @@ All notable changes to this project will be documented here.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### ✨ Enhancements
+
+- Create `useCursorPosition` hook.
+
 ## 1.5.1 (June 19, 2020)
 
 ### 🐛 Bug fixes